Add (basic for now) closures and change everything else to support them
This commit is very big. The main goal was to support closures, but doing so correctly and with low overhead required: changing the way values are stored on the stack, changing the way functions are called, adding a runtime type system, adding a constant pool, adding an additional metadata-gathering compilation pass, improving support for lists, and much more.

Along the way I also got distracted and decided to overhaul the bridge system to make it support both runtime compilation and pre-compilation via a separate process, wherein the bindings need to be serialized before being used. This also allows for lower overhead at runtime. Finally, this commit also adds a basic runtime overrides system to the existing wrapper class concept, which will be the basis for selective runtime code-push in the future.